add build time Currently getStaticProps is "temporarily" disabled for _app.js, making fetching common properties across all pages at *build time* not trivial.
This approach fetches the build time for each page individually, so although Vercel server is very fast, one can expect the time to be different by a few seconds.

If this actually works that way, I will see if I can make a property in _app.js to gather all buildTime's together to use a uniform buildTime.